chore: rename middleware to awsAuthMiddleware
