+++ packages/middleware-signing/src/awsAuthMiddleware.ts
