+++ packages/middleware-signing/src/getAwsAuthPlugin.ts
@@ -1,7 +1,7 @@
 import { Pluggable } from "@aws-sdk/types";
 
+import { awsAuthMiddleware, awsAuthMiddlewareOptions } from "./awsAuthMiddleware";
 import { AwsAuthResolvedConfig } from "./configurations";
-import { awsAuthMiddleware, awsAuthMiddlewareOptions } from "./middleware";
 
 export const getAwsAuthPlugin = (options: AwsAuthResolvedConfig): Pluggable<any, any> => ({
   applyToStack: (clientStack) => {
