+++ packages/middleware-signing/src/index.ts
@@ -1,4 +1,4 @@
+export * from "./awsAuthMiddleware";
 export * from "./configurations";
 export * from "./getAwsAuthPlugin";
 export * from "./getSigV4AuthPlugin";
-export * from "./middleware";
